register

+++ yoga/RL_LoginViewController.m
@@ -99,10 +99,9 @@
 }
 -(void)btnClick:(UIButton *)btn
 {
-    NSLog(@"%d",btn.tag);
+    NSLog(@"%ld",(long)btn.tag);
     if (btn.tag == 2) {
-        NSDictionary *parameter = [NSDictionary dictionaryWithObjectsAndKeys:_account.text,@"username",
-                                                                        _passWord.text,@"password",nil];
+        NSDictionary *parameter = [NSDictionary dictionaryWithObjectsAndKeys:_account.text,@"username",_passWord.text,@"password",nil];
         AFHTTPRequestOperationManager *manager = [[AFHTTPRequestOperationManager alloc] init];
         manager.responseSerializer.acceptableContentTypes = [NSSet setWithObject:@"text/html"];
